make pndman_package_get_embedded_png return number of bytes copied

+++ include/pndman.h
@@ -469,8 +469,8 @@ PNDMANAPI int pndman_package_crawl_single_package(int full_crawl,
  * you need to provide your buffer and it's size.
  * if this functions returns 0, your buffer is filled
  * with the png data.
- * returns 0 on succes, -1 on failure */
-PNDMANAPI int pndman_package_get_embedded_png(
+ * returns number of bytes copied on success, 0 on failure */
+PNDMANAPI size_t pndman_package_get_embedded_png(
       pndman_package *pnd, char *buffer, size_t buflen);
 
 /* \brief initialize package handle
